chore(form): refactor and move useForm to composable
