+++ packages/ui/src/components/va-counter/VaCounter.vue
@@ -96,7 +96,7 @@ import pick from 'lodash/pick'
 import { safeCSSLength } from '../../utils/css'
 import {
   useComponentPresetProp,
-  useFormProps,
+  useFormFieldProps,
   useEmitProxy,
   useFocus, useFocusEmits,
   useStateful, useStatefulProps,
@@ -125,7 +125,7 @@ export default defineComponent({
   components: { VaInputWrapper, VaButton },
 
   props: {
-    ...useFormProps,
+    ...useFormFieldProps,
     ...useStatefulProps,
     ...useComponentPresetProp,
     // input
