+++ packages/ui/src/composables/tests/useFormField.spec.ts
@@ -1,4 +1,4 @@
-import { useForm } from '../useForm'
+import { useFormField } from '../useFormField'
 import { describe, it, expect } from 'vitest'
 
 describe('useForm', () => {
@@ -8,10 +8,9 @@ describe('useForm', () => {
     [{ disabled: true, readonly: false } , 'va-test'        , { 'va-test--disabled': true }                           ],
     [{ disabled: true, readonly: false } , ''               , { '--disabled': true }                                  ],
     [{ disabled: true, readonly: true }  , 'va-test'        , { 'va-test--disabled': true, 'va-test--readonly': true }],
-    [{}                                  , 'va-test'        , {}                                                      ],
     /* eslint-enable */
   ])('props %s & prefix %s should be %s', (props, prefix, expected) => {
-    const { computedClasses } = useForm(prefix, props)
+    const { computedClasses } = useFormField(prefix, props)
     expect(computedClasses.asObject.value).toMatchObject(expected)
   })
 })
