+++ packages/ui/src/composables/useForm/consts.ts
@@ -1,0 +1,4 @@
+import { InjectionKey } from 'vue'
+import { type createFormContext } from './useFormParent'
+
+export const FormServiceKey: InjectionKey<ReturnType<(typeof createFormContext)>> = Symbol('FormService')
