+++ packages/ui/src/composables/useForm/types.ts
@@ -1,0 +1,10 @@
+export type FormFiled = {
+  name?: string;
+  value: unknown;
+  isValid: boolean;
+  errorMessages: string[];
+  validate: () => boolean;
+  reset: () => void;
+  resetValidation: () => void;
+  focus: () => void;
+}
